+++ google2/CHANGELOG.md
@@ -10,6 +10,7 @@ SBT Dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0
 * Added ability for getBucket to log at WARN level instead of ERROR
 * Updated GCE, Dataproc, and GKE interpreters to use `IO.blocking` combinator for blocking API calls through 
 the Google SDK.
+* Add `patchReplicas` to `KubernetesService`
 
 ### Dependency upgrades
 | Dependency                    |     Old Version      |          New Version |